完成魔咒效果组件: 易碎

+++ wakame-mixin/src/main/kotlin/cc/mewcraft/wakame/enchantment2/component/Fragile.kt
@@ -4,7 +4,7 @@ import com.github.quillraven.fleks.Component
 import com.github.quillraven.fleks.ComponentType
 
 class Fragile(
-    val multiplier: Int,
+    val multiplier: Float,
 ) : Component<Fragile> {
 
     companion object : ComponentType<Fragile>()
